pyproject.toml
[tool.poetry]
name = "reporting-tool"
version = "0.1.0"
description = ""
authors = ["DatAction-Arne <96430642+DatAction-Arne@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "reporting_tool"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pandas = "^2.0.0"
numpy = "^1.24.2"
codetiming = "^1.4.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.4"
mysql-connector-python = "^8.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
